requirements.txt
httpx==0.28.1
PyYAML==6.0.2
base58==2.1.1
pandas==2.2.3
pyarrow==17.0.0
beautifulsoup4==4.12.3
pymongo==4.10.1
matplotlib==3.9.2
